* WARNING: I wasn't able to see that this compiles all the way through -- it should, but there is a chance it might not.. * fix for maketa under sanitize=address -- was actually using overwritten strings * configure script now has sanitize option to enable it -- under clang * char* conversion for string wasn't using proper converter -- is now * MASSIVE fixes for override keyword -- needed to prevent warnings in c++11 mode -- still needs a lot more, especially css_c_ptr_types -- css guys were brutal.

+++ src/temt/ta_core/TypeDef.cpp
@@ -2607,7 +2607,7 @@ void TypeDef::CopyFromSameType(void* trg_base, void* src_base,
 	*((taAtomicInt*)trg_base) = *((taAtomicInt*)src_base);
       }
       else if(DerivesFrom(TA_taBasicAtomicInt)) {
-	*((taBasicAtomicInt*)trg_base) = *((taBasicAtomicInt*)src_base);
+	*((taBasicAtomicInt*)trg_base) = (int)*((taBasicAtomicInt*)src_base);
       }
 #endif
     }
